+++ coco/src/App.js
@@ -77,12 +77,13 @@ import { Login } from "./components/Login/Login";
 import { Problems } from "./components/Problems/Problems";
 import { PBD } from "./components/ProblemDetail/PBD";
 import { Result } from "./components/Result/Result";
-import { Manage } from "./components/Manage/Manage";
+import { ManageLogin } from "./components/Manage/ManageLogin";
 import { Board } from "./components/Board/Board";
 import { BoardDetail } from "./components/Board/BoardDetail/BoardDetail";
 import { WriteGeul } from "./components/Board/WriteGeul/WriteGuel";
 import { StatusList } from "./components/StatusList/StatusList";
 import { MyPage } from "./components/MyPage/MyPage";
+import { Manage } from "./components/Manage/Manage";
 
 function App() {
   /* 토큰이 유효하지 않으면 토큰 초기화*/
@@ -113,7 +114,8 @@ function App() {
           <Route path="/board" element={<Board />} />
           <Route path="/board/:id" element={<BoardDetail />} />
           <Route path="/mypage/:id" element={<MyPage />} />
-          <Route path="/manage" element={<Manage />} />
+          <Route path="/manage" element={<ManageLogin />} />
+          <Route path="/manageSuccess" element={<Manage />} />
           <Route
             path="/write"
             element={
