+++ coco/src/components/Manage/Manage.css
@@ -1,3 +1,152 @@
+/*-------------로그인 섹션------------------*/
+.manageLogin {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  gap: 50px;
+  background-color: #222222;
+}
+
+.manageLogin h2 {
+  font-family: "Pretendard-Regular";
+  font-weight: 600;
+  font-size: 2.2rem;
+  color: skyblue;
+  width: 330px;
+  text-align: center;
+  border: 1px solid rgb(176, 229, 252);
+  padding: 10px;
+  margin-bottom: 70px;
+}
+
+.form__group {
+  position: relative;
+  padding: 15px 0 0;
+}
+.form__field {
+  font-family: "Pretendard-Regular";
+  width: 100%;
+  width: 330px;
+  border: 0;
+  border-bottom: 2px solid skyblue;
+  outline: 0;
+  font-size: 1.5rem;
+  color: #fff;
+  padding: 6px 0;
+  background: transparent;
+  transition: border-color 0.2s;
+}
+.form__field::placeholder {
+  color: transparent;
+}
+.form__field:placeholder-shown ~ .form__label {
+  font-size: 1.2rem;
+  cursor: text;
+  top: 20px;
+}
+.form__label {
+  position: absolute;
+  top: 0;
+  display: flex;
+  transition: 0.2s;
+  color: #9b9b9b;
+}
+.form__field:focus {
+  padding-bottom: 6px;
+  font-weight: 600;
+  border-width: 3px;
+  border-image: linear-gradient(to right, #71cafd, #3872ef);
+  border-image-slice: 1;
+}
+.form__field:focus ~ .form__label {
+  position: absolute;
+  top: 0;
+  display: block;
+  transition: 0.1s;
+  font-size: 1.1rem;
+  color: #3872ef;
+  font-weight: 600;
+}
+
+.form__field:required,
+.form__field:invalid {
+  box-shadow: none;
+}
+
+.manageLogin div[class*="box"] {
+  margin-top: 30px;
+  width: 100%;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.manageLogin .btn {
+  background-color: rgba(34, 34, 34, 0.1);
+  font-family: "Pretendard-Regular";
+  font-weight: 500;
+  font-size: 1.3rem;
+  line-height: 45px;
+  letter-spacing: 2px;
+  height: 60px;
+  text-align: center;
+  width: 330px;
+  cursor: pointer;
+}
+
+.manageLogin .btn-one {
+  color: skyblue;
+  transition: all 0.4s;
+  position: relative;
+}
+.manageLogin .btn-one span {
+  transition: all 0.3s;
+}
+.manageLogin .btn-one::before {
+  content: "";
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 1;
+  opacity: 0;
+  transition: all 0.3s;
+  border-top-width: 1px;
+  border-bottom-width: 1px;
+  border-top-style: solid;
+  border-bottom-style: solid;
+  border-top-color: rgba(255, 255, 255, 0.5);
+  border-bottom-color: rgba(255, 255, 255, 0.5);
+  transform: scale(0.1, 1);
+}
+.manageLogin .btn-one:hover span {
+  letter-spacing: 5px;
+}
+.manageLogin .btn-one:hover::before {
+  opacity: 1;
+  transform: scale(1, 1);
+}
+.manageLogin .btn-one::after {
+  content: "";
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 1;
+  transition: all 0.3s;
+  background-color: rgba(34, 34, 34, 0.1);
+}
+.manageLogin .btn-one:hover::after {
+  opacity: 0;
+  transform: scale(0.1, 1);
+}
+
+/*-------------로그인 섹션------------------*/
 .manage {
   display: flex;
   flex-flow: column;
