+++ coco/src/components/Manage/manageComponents/Notice.jsx
@@ -1,0 +1,6 @@
+import "../Manage.css";
+import React from "react";
+
+export const Notice = () => {
+  return <div>Notice</div>;
+};
